refactor: Move xudt examples to signle package
