--- examples/rgbpp/xudt/0-token-info.ts
@@ -1,7 +1,0 @@
-import { RgbppTokenInfo } from '@rgbpp-sdk/ckb';
-
-export const XUDT_TOKEN_INFO: RgbppTokenInfo = {
-  decimal: 8,
-  name: 'XUDT Test Token',
-  symbol: 'XTT',
-};
