+++ examples/xudt/.env.example
@@ -1,0 +1,11 @@
+# The secp256k1 private key whose format is 32bytes hex string with 0x prefix
+CKB_SECP256K1_PRIVATE_KEY='0x-private-key'
+
+# Ture for CKB Mainnet and false for CKB Testnet, the default value is false
+IS_MAINNET='false'
+
+# CKB node url which should be matched with IS_MAINNET
+CKB_NODE_URL='https://testnet.ckb.dev/rpc'
+
+# CKB indexer url which should be matched with IS_MAINNET
+CKB_INDEXER_URL='https://testnet.ckb.dev/indexer'
