+++ examples/xudt/README.md
@@ -1,0 +1,17 @@
+# xUDT on CKB Examples
+
+The examples for xUDT issuance, mint and transfer on CKB
+
+### Issue xUDT on CKB
+
+```shell
+npx ts-node examples/rgbpp/xudt/1-issue-xudt.ts 
+```
+
+### Mint/Transfer xUDT on CKB
+
+You can use this command to mint or transfer xUDT assets
+
+```shell
+npx ts-node examples/rgbpp/xudt/2-transfer-xudt.ts 
+```
