+++ examples/xudt/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "xudt-examples",
+  "version": "0.1.0",
+  "description": "Examples used for xUDT on CKB assets issuance and transfer",
+  "private": true,
+  "type": "commonjs",
+  "scripts": {
+    "format": "prettier --write '**/*.ts'",
+    "lint": "tsc && eslint . && prettier --check '**/*.ts'",
+    "lint:fix": "tsc && eslint --fix --ext .ts . && prettier --write '**/*.ts'"
+  },
+  "dependencies": {
+    "@nervosnetwork/ckb-sdk-utils": "^0.109.1",
+    "@rgbpp-sdk/ckb": "workspace:^"
+  }
+}
